// src/types/carTypes/ICar.ts
// Generated by https://quicktype.io

export interface ICar {
    _created: number;
    _data_type: string;
    _is_deleted: boolean;
    _modified: number;
    _self_link: string;
    _user: string;
    _uuid: string;
    fuel: string;
    km_driven: number;
    name: string;
    owner: string;
    seller_type: string;
    selling_price: number;
    transmission: string;
    image: string;
    year: number;
}